Moved all DAO calls from employee controllers to employeeService, adding a service layer in between.
